Add files via upload
Data Set Information:

Features are computed from a digitized image of a fine needle aspirate (FNA) of a breast mass. They describe characteristics of the cell nuclei present in the image. A few of the images can be found at [Web Link]

by using the given independent features we have to predict the new record class whether it is abnormal or normal